+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Standup read failure must refuse, not render an empty standup ([pm-slack-standup-011l](https://github.com/unbraind/pm-slack-standup/blob/main/.agents/pm/issues/pm-slack-standup-011l.toon))
+
 ## 2026.8.14 - 2026-08-14
 
 ### Fixed